style: 채팅 관련 페이지를 ChatPage 폴더 내부로 이동 (#444)
